Moving DungeonRoom to separate class
