+++ Assets/Scripts/Terrain/Generator/Structure/Dungeon/DungeonRoom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 049075b0773f4e0998ebd397e1b3eaf3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
