+++ Assets/Scripts/Terrain/Generator/Structure/Dungeon/DungeonRoomTree.cs
@@ -42,25 +42,6 @@ namespace Terrain.Generator.Structure.Dungeon
         //     
         // }
     }
-    
-    public struct DungeonRoom : IDungeonRoom
-    {
-        public AABB2D Rect { get; set;}
-
-        public float Area
-        {
-            get
-            {
-                float2 d = Rect.max - Rect.min;
-                return d.x * d.y;
-            }
-        }
-
-        public DungeonRoom(int2 pos, int2 size)
-        {
-            Rect = new AABB2D(pos, pos + size);
-        }
-    }
 
     public interface IDungeonRoom
     {
